+++ R/getPredictorsOpenMx.R
@@ -12,7 +12,7 @@ getPredictorsOpenMx <- function(mxmodel, dataset, covariates)
       model.ids[i] <- which(cmp);
     }
     all.ids <- 1:length(names(dataset))
-    cvid <- sets::as.set(all.ids)-sets::as.set(model.ids) 
+    cvid <- all.ids[!all.ids %in% model.ids]
     if (length(cvid)==0) {
       ui_stop("Error. No predictors contained in dataset!")
     }
@@ -23,7 +23,7 @@ getPredictorsOpenMx <- function(mxmodel, dataset, covariates)
     all.ids <- 1:length(names(dataset))
     covariate.ids <- sapply(covariates, function(cv) { which(cv==names(dataset))} )
     
-    modid <- sets::as.set(all.ids)-sets::as.set(covariate.ids) 
+    modid <- all.ids[!all.ids %in% covariate.ids]
     if (length(modid)==0) {
       ui_stop("No covariates contained in dataset!")
     }